+++ test/unit/fuzz-tools.test.js
@@ -134,6 +134,7 @@ import * as warfarinV133 from '../../lib/warfarin-v133.js';
 import * as oncV134 from '../../lib/onc-v134.js';
 import * as lymphomaV135 from '../../lib/lymphoma-v135.js';
 import * as endoV136 from '../../lib/endo-v136.js';
+import * as idV137 from '../../lib/id-v137.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
@@ -213,6 +214,7 @@ const MODULES = {
   'onc-v134.js': oncV134,
   'lymphoma-v135.js': lymphomaV135,
   'endo-v136.js': endoV136,
+  'id-v137.js': idV137,
   'ems-v149.js': emsV149,
 };